application/requirements.txt
amqp==2.4.2
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
backcall==0.1.0
billiard==3.6.0.0
boto==2.49.0
boto3==1.9.215
botocore==1.12.215
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
Django==2.1.7
django-celery-results==1.0.4
django-cors-headers==2.5.2
django-minio-storage==0.2.2
djangorestframework==3.9.2
docutils==0.15.2
ExifRead==2.1.2
feedparser==5.2.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.15
jedi==0.13.3
jmespath==0.9.4
kombu==4.5.0
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
minio==4.0.20
mysqlclient==1.4.2.post1
parso==0.3.4
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.1
six==1.12.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.1
